+++ Data/StableMasters.lua
@@ -22,9 +22,11 @@ TF_STABLE_MASTERS = {
     { zone = TF_ZONE.THUNDER_BLUFF, x = 0.4511, y = 0.6022, name = "Bulrug", faction = TF_FACTION.HORDE },
 	{ zone = TF_ZONE.ARATHI_HIGHLANDS, x = 0.7393, y = 0.3313, name = "Tharlidun", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.BADLANDS, x = 0.0367, y = 0.4759, name = "Greth", faction = TF_FACTION.HORDE },
+    { zone = TF_ZONE.ELWYNN_FOREST, x = 0.4285, y = 0.6594, name = "Erma", faction = TF_FACTION.ALLIANCE },
     { zone = TF_ZONE.STORMWIND_CITY, x = 0.6147, y = 0.1716, name = "Jenova Stoneshield", faction = TF_FACTION.ALLIANCE },
     { zone = TF_ZONE.STORMWIND_CITY, x = 0.2959, y = 0.5122, name = "Sylista", faction = TF_FACTION.ALLIANCE },
     { zone = TF_ZONE.THOUSAND_NEEDLES, x = 0.4577, y = 0.5106, name = "Awenasa", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.FERALAS, x = 0.7449, y = 0.4327, name = "Shyrka Wolfrunner", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.TANARIS, x = 0.5225, y = 0.2800, name = "Laziphus" },
+    { zone = TF_ZONE.WESTFALL, x = 0.5294, y = 0.5307, name = "Kirk Maxwell", faction = TF_FACTION.ALLIANCE },
 }