+++ changelog.md
@@ -1,4 +1,8 @@
-### v1.0.2 (Sept. ?, 2019)
+### v1.0.2 (Sept. 13, 2019)
 * Added Thousand Needles NPCs
 * Added Camp Mojache, Feralas NPCs
 * Added Tanaris NPCs
+* Added Elwynn Forest NPCs
+* Added Westfall NPCs
+* Added a few missing Stormwind NPCs
+* Minor locale fixes